model/requirements.txt
google-cloud-aiplatform
tensorflow
keras-cv
Pillow
